figma-plugin/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
langgraph>=0.0.45
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.0.4
langchain-google-genai>=0.0.8
langchain-groq>=0.0.3
replicate>=0.15.0
firecrawl-py>=0.0.8
python-multipart>=0.0.6
aiofiles>=23.2.1
supabase>=2.3.0
python-dotenv>=1.0.0